Resolve spotbugs, copyright, and one change to go back to Java 21
