+++ extensions/openapi-generator/modules/openapi-generator/pom.xml
@@ -38,6 +38,7 @@
 
     <properties>
         <helidon.services.skip>true</helidon.services.skip>
+        <spotbugs.exclude>etc/spotbugs/exclude.xml</spotbugs.exclude>
     </properties>
 
     <dependencies>
