+++ extensions/openapi-generator/modules/openapi-generator/src/main/java/io/helidon/openapi/generator/JavaStringLiterals.java
@@ -1,0 +1,49 @@
+/*
+ * Copyright (c) 2026 Oracle and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.helidon.openapi.generator;
+
+final class JavaStringLiterals {
+
+    private JavaStringLiterals() {
+    }
+
+    static String toJavaStringLiteral(String value) {
+        StringBuilder result = new StringBuilder("\"");
+        for (int i = 0; i < value.length(); i++) {
+            char ch = value.charAt(i);
+            switch (ch) {
+                case '\\' -> result.append("\\\\");
+                case '"' -> result.append("\\\"");
+                case '\n' -> result.append("\\n");
+                case '\r' -> result.append("\\r");
+                case '\t' -> result.append("\\t");
+                case '\b' -> result.append("\\b");
+                case '\f' -> result.append("\\f");
+                default -> {
+                    if (Character.isISOControl(ch)) {
+                        result.append("\\u");
+                        String hex = Integer.toHexString(ch);
+                        result.append("0".repeat(4 - hex.length())).append(hex);
+                    } else {
+                        result.append(ch);
+                    }
+                }
+            }
+        }
+        return result.append('"').toString();
+    }
+}
